pyproject.toml
[project]
name = "zapi"
version = "0.1.0"
description = "API collections for study"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "markdown-it-py>=3.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.6",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "alembic>=1.12.1",
]

[[tool.uv.index]]
# Optional name for the index.
name = "aliyun"
# Required URL for the index.
url = "https://mirrors.aliyun.com/pypi/simple/"
# exclude PyPI
# default = true
